fix: Convert float position calculations to int in Enemy.shoot_at_angle
- Fix type error by converting x, y coordinates to int when creating EnemyShot

- Prevent float values from being passed to int parameters

+++ src/components/enemy.py
@@ -177,8 +177,8 @@ class Enemy(Sprite):
         """
         s = EnemyShot(
             self.game_state,
-            self.x + (self.w / 2) + offset_x,
-            self.y + (self.h / 2) + offset_y,
+            int(self.x + (self.w / 2) + offset_x),
+            int(self.y + (self.h / 2) + offset_y),
             px.cos(degrees) * speed,
             px.sin(degrees) * speed,
             delay,
